+++ CHANGELOG.md
@@ -74,3 +74,4 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Changed
 - Rename accept and reject bridgehead state buttons (authorize/revoke)
+- Rename reject project bridgehead user state as block